+++ graph/count_theorem_definition_label.py
@@ -2,7 +2,7 @@ import os
 import glob
 import pprint
 import json
-from invest_diff_ver_article import extract_theorem_definition_and_label
+import re
 
 def make_mizar_file_path(mml_version):
     cwd = os.getcwd()
@@ -82,6 +82,68 @@ def calc_average_of_theorems_and_definitions(miz_file2theorems_and_definitions):
     for num in total_theorem_or_definition2number.values():
         total += num
     return total / total_miz_file
+
+def extract_theorem_definition_and_label(article):
+    # 単語、改行、::、;で区切ってファイルの内容を取得
+    file_words = re.findall(r"\w+:*|\n|::|;|\.=", article)
+    is_comment = False
+    is_theorem_or_label = False
+    theorem_and_label = list()
+
+    # mizファイルからbyで引用するtheorem,labelを取得する
+    for word in file_words:
+        # コメント行の場合
+        if word == "::" and not is_comment:
+            is_comment = True
+            continue
+        # コメント行の終了
+        if re.search(r"\n", word) and is_comment:
+            is_comment = False
+        # byの検索
+        if word == "by" and not is_comment:
+            is_theorem_or_label = True
+            continue
+        # byで引用する構文の終了
+        if (re.search(r";", word) or re.search(r"\.=", word)) and is_theorem_or_label:
+            is_theorem_or_label = False
+        # byで引用するtheorem, labelを取得
+        if is_theorem_or_label and word != "\n":
+            if theorem_and_label:
+                if re.search(r":$|def$", theorem_and_label[-1]):
+                    theorem_and_label[-1] = theorem_and_label[-1] + word
+                    continue
+                elif re.match(r"\d+", word):
+                    if re.search(r":def\d+$|:\d+$", theorem_and_label[-1]):
+                        new_theorem = theorem_and_label[-1]
+                        theorem_and_label.append(re.sub(r"\d+$", word, new_theorem))
+                    continue
+            theorem_and_label.append(word)
+        
+    return theorem_and_label
+
+def get_numnber_of_proof(old_article, new_article):
+    return count_proof(old_article), count_proof(new_article)
+
+def count_proof(article):
+    # 単語、改行、::、;で区切ってファイルの内容を取得
+    file_words = re.findall(r"\w+|\n|::", article)
+    is_comment = False
+    proof_counter = 0
+
+    # mizファイルからbyで引用するtheorem,labelを取得する
+    for word in file_words:
+        # コメント行の場合
+        if word == "::" and not is_comment:
+            is_comment = True
+            continue
+        # コメント行の終了
+        if re.search(r"\n", word) and is_comment:
+            is_comment = False
+        # コメント外でproofが出現
+        if not is_comment and word == "proof":
+            proof_counter += 1
+
+    return proof_counter
 
 def main(mml_version):
     miz_file2theorem_or_definition2number = dict()
